Conflicts:
	pyon/core/bootstrap.py
	pyon/ion/exchange.py

+++ pyon/net/channel.py
@@ -751,7 +751,7 @@ class ListenChannel(RecvChannel):
         self._on_stop_consume(fsm)
 
     @contextmanager
-    def accept(self):
+    def accept(self, timeout=None):
         """
         Context manager method to accept new connections for listening endpoints.
 
@@ -768,7 +768,7 @@ class ListenChannel(RecvChannel):
                     - recv() returns messages in its gqueue, endpoint should ack
         """
         #        self._ensure_amq_chan()
-        m = self.recv()
+        m = self.recv(timeout=timeout)
         ch = self._create_accepted_channel(self._amq_chan, m)
         ch._recv_queue.put(m)       # prime our recieved message here, should be acked by EP layer
 
@@ -779,7 +779,7 @@ class ListenChannel(RecvChannel):
 
 class SubscriberChannel(ListenChannel):
     def close_impl(self):
-        if not self._queue_auto_delete and self._recv_name and self._recv_name.queue.startswith("amq.gen-"):
+        if not self._queue_auto_delete and self._recv_name and self._recv_name.queue.startswith("amq.gen-") and self._transport is AMQPTransport.get_instance():
             log.debug("Anonymous Subscriber detected, deleting queue (%s)", self._recv_name)
             self._destroy_queue()
